feat: adiciona o frontend

+++ backend/Entities/Herois.cs
@@ -6,7 +6,7 @@ public class Herois
     public string Nome { get; set; }
     public string NomeHeroi { get; set; }
     public DateTime DataNascimento { get; set; }
-    public float Altura { get; set; }
-    public float Peso { get; set; }
+    public double Altura { get; set; }
+    public double Peso { get; set; }
     public ICollection<Superpoderes> Superpoderes { get; set; }
 }
